✨feat: TeamIssue 창에서 dataprovider을 사용하여 api를 호출해보기
