+++ tsconfig.paths.json
@@ -10,7 +10,8 @@
       "@models/*": ["src/application/models/*"],
       "@stores/*": ["src/application/stores/*"],
       "@hooks/*": ["src/application/hooks/*"],
-      "@api/*": ["src/infrastructure/api/*"]
+      "@api/*": ["src/infrastructure/api/*"],
+      "@provider/*": ["src/infrastructure/provider/*"]
     }
   }
 }
